+++ watchdog_bringup/scripts/drivers/ids_check.sh
@@ -2,16 +2,16 @@
 # this bash script adheres to linux exit codes
 # see https://tldp.org/LDP/abs/html/exitcodes.html for further information
 
-echo "[BASH] checking sensor - FAILURE"
+echo "[BASH - IDS] checking sensor - IDS Camera"
 exit 1
 
 IDS_STATUS="$(systemctl status ueyeusbdrc)"
-echo "$IDS_STATUS"
+echo "[BASH - IDS] status: ${IDS_STATUS}"
 
 case $IDS_STATUS in
   "")
     echo "[BASH - IDS] Service cannot be found"
-    exit 0
+    exit 1
     ;;
 esac